+++ src/Eloquent/AbstractAdapter.php
@@ -48,6 +48,7 @@ abstract class AbstractAdapter extends AbstractResourceAdapter
 {
 
     use FindsManyResources,
+        Concerns\DeserializesAttributes,
         Concerns\IncludesModels;
 
     /**
@@ -70,29 +71,6 @@ abstract class AbstractAdapter extends AbstractResourceAdapter
     protected $primaryKey;
 
     /**
-     * Mapping of JSON API attribute field names to model keys.
-     *
-     * By default, JSON API attribute fields will automatically be converted to the
-     * underscored or camel cased equivalent for the model key. E.g. if the model
-     * uses snake case, the JSON API field `published-at` will be converted
-     * to `published_at`. If the model does not use snake case, it will be converted
-     * to `publishedAt`.
-     *
-     * For any fields that do not directly convert to model keys, you can list them
-     * here. For example, if the JSON API field `published-at` needed to map to the
-     * `published_date` model key, then it can be listed as follows:
-     *
-     * ```php
-     * protected $attributes = [
-     *   'published-at' => 'published_date',
-     * ];
-     * ```
-     *
-     * @var array
-     */
-    protected $attributes = [];
-
-    /**
      * JSON API relationship field names to hydrate.
      *
      * Lists the JSON API relationship field names that can be hydrated (modified).
@@ -100,16 +78,6 @@ abstract class AbstractAdapter extends AbstractResourceAdapter
      * @var string[]
      */
     protected $relationships = [];
-
-    /**
-     * The resource attributes that are dates.
-     *
-     * If an array, a list of JSON API attribute fields that should be cast to dates.
-     * If `null`, the list will be calculated using `Model::getDates()`
-     *
-     * @var string[]|null
-     */
-    protected $dates = null;
 
     /**
      * The filter param for a find-many request.
@@ -434,71 +402,7 @@ abstract class AbstractAdapter extends AbstractResourceAdapter
      */
     protected function keyForAttribute($resourceKey, Model $model)
     {
-        return $this->modelKeyForField($resourceKey);
-    }
-
-    /**
-     * Convert a JSON API resource field name to a model key.
-     *
-     * @param $field
-     * @return string
-     */
-    protected function modelKeyForField($field)
-    {
-        if (isset($this->attributes[$field])) {
-            return $this->attributes[$field];
-        }
-
-        $model = $this->model;
-        $key = $model::$snakeAttributes ? Str::underscore($field) : Str::camelize($field);
-
-        return $this->attributes[$field] = $key;
-    }
-
-    /**
-     * Deserialize a value obtained from the resource's attributes.
-     *
-     * @param $value
-     *      the value that the client provided.
-     * @param $resourceKey
-     *      the attribute key for the value
-     * @param Model $record
-     * @return Carbon|null
-     */
-    protected function deserializeAttribute($value, $resourceKey, $record)
-    {
-        if ($this->isDateAttribute($resourceKey, $record)) {
-            return $this->deserializeDate($value);
-        }
-
-        return $value;
-    }
-
-    /**
-     * Convert a JSON date into a PHP date time object.
-     *
-     * @param $value
-     * @return Carbon|null
-     */
-    protected function deserializeDate($value)
-    {
-        return !is_null($value) ? new Carbon($value) : null;
-    }
-
-    /**
-     * Is this resource key a date attribute?
-     *
-     * @param $field
-     * @param Model $record
-     * @return bool
-     */
-    protected function isDateAttribute($field, $record)
-    {
-        if (is_null($this->dates)) {
-            return in_array($this->modelKeyForField($field), $record->getDates(), true);
-        }
-
-        return in_array($field, $this->dates, true);
+        return $this->modelKeyForField($resourceKey, $model);
     }
 
     /**